+++ test/client/utils.js
@@ -29,11 +29,7 @@ function verifyRequestCall(stubArgs, method, uri, body, config) {
 }
 
 function sleep(ms) {
-    return new Promise((resolve) => {
-        setTimeout(() => {
-            return resolve()
-        }, ms);
-    })
+    return new Promise((resolve) => setTimeout(resolve, ms))
 }
 
 module.exports = {